Refactor moving SingleChoiceButton and MultipleChoiceButton
