+++ shared/src/commonMain/kotlin/ui/views/iso/verifier/VerifierCustomSelectionView.kt
@@ -43,6 +43,8 @@ import ui.composables.Logo
 import ui.composables.ScreenHeading
 import ui.composables.buttons.NavigateUpButton
 import ui.viewmodels.iso.verifier.VerifierViewModel
+import ui.views.iso.common.MultipleChoiceButton
+import ui.views.iso.common.SingleChoiceButton
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
