+++ shared/src/commonMain/kotlin/ui/views/iso/verifier/VerifierDocumentSelectionView.kt
@@ -41,6 +41,7 @@ import org.jetbrains.compose.resources.stringResource
 import ui.composables.Logo
 import ui.composables.ScreenHeading
 import ui.viewmodels.iso.verifier.VerifierViewModel
+import ui.views.iso.common.SingleChoiceButton
 import ui.views.iso.verifier.requests.AVRequests
 import ui.views.iso.verifier.requests.HIIDRequest
 import ui.views.iso.verifier.requests.MDLRequests
